Refactored BookController to thin it out and refactor some of the functions to apply IOC. Added some split query statements on a few queries.
